feat: helpers methods
